Add support for SystemTable's to delegate split processing to connectors
