+++ core/trino-main/src/main/java/io/trino/connector/DefaultCatalogFactory.java
@@ -19,8 +19,7 @@ import io.airlift.configuration.secrets.SecretsResolver;
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.trace.Tracer;
 import io.trino.connector.informationschema.InformationSchemaConnector;
-import io.trino.connector.system.CoordinatorSystemTablesProvider;
-import io.trino.connector.system.StaticSystemTablesProvider;
+import io.trino.connector.system.DefaultSystemTablesProvider;
 import io.trino.connector.system.SystemConnector;
 import io.trino.connector.system.SystemTablesProvider;
 import io.trino.execution.scheduler.NodeSchedulerConfig;
@@ -156,17 +155,11 @@ public class DefaultCatalogFactory
                         accessControl,
                         maxPrefetchedInformationSchemaPrefixes));
 
-        SystemTablesProvider systemTablesProvider;
-        if (currentNode.isCoordinator()) {
-            systemTablesProvider = new CoordinatorSystemTablesProvider(
-                    transactionManager,
-                    metadata,
-                    catalogHandle.getCatalogName().toString(),
-                    new StaticSystemTablesProvider(catalogConnector.getSystemTables()));
-        }
-        else {
-            systemTablesProvider = new StaticSystemTablesProvider(catalogConnector.getSystemTables());
-        }
+        SystemTablesProvider systemTablesProvider = new DefaultSystemTablesProvider(
+                transactionManager,
+                metadata,
+                catalogHandle.getCatalogName().toString(),
+                catalogConnector.getSystemTables());
 
         ConnectorServices systemConnector = new ConnectorServices(
                 tracer,
@@ -177,7 +170,7 @@ public class DefaultCatalogFactory
                         systemTablesProvider,
                         transactionId -> transactionManager.getConnectorTransaction(transactionId, catalogHandle),
                         accessControl,
-                        catalogHandle.getCatalogName().toString()));
+                        catalogHandle.getCatalogName().toString(), catalogConnector.getPageSourceProviderFactory()));
 
         return new CatalogConnector(
                 catalogHandle,
