+++ core/trino-main/src/main/java/io/trino/connector/system/DefaultSystemTablesProvider.java
@@ -18,67 +18,67 @@ import io.trino.FullConnectorSession;
 import io.trino.Session;
 import io.trino.metadata.Metadata;
 import io.trino.metadata.QualifiedObjectName;
-import io.trino.spi.TrinoException;
 import io.trino.spi.connector.ConnectorSession;
 import io.trino.spi.connector.SchemaTableName;
 import io.trino.spi.connector.SystemTable;
 import io.trino.transaction.TransactionManager;
 
+import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 
-import static io.trino.spi.StandardErrorCode.GENERIC_INTERNAL_ERROR;
-import static io.trino.spi.connector.SystemTable.Distribution.SINGLE_COORDINATOR;
+import static com.google.common.collect.ImmutableMap.toImmutableMap;
 import static java.util.Objects.requireNonNull;
+import static java.util.function.Function.identity;
 
-public class CoordinatorSystemTablesProvider
+public class DefaultSystemTablesProvider
         implements SystemTablesProvider
 {
     private final TransactionManager transactionManager;
     private final Metadata metadata;
     private final String catalogName;
-    private final StaticSystemTablesProvider staticProvider;
+    private final Set<SystemTable> systemTables;
+    private final Map<SchemaTableName, SystemTable> systemTablesMap;
 
-    public CoordinatorSystemTablesProvider(TransactionManager transactionManager, Metadata metadata, String catalogName, StaticSystemTablesProvider staticProvider)
+    public DefaultSystemTablesProvider(TransactionManager transactionManager, Metadata metadata, String catalogName, Set<SystemTable> systemTables)
     {
         this.transactionManager = requireNonNull(transactionManager, "transactionManager is null");
         this.metadata = requireNonNull(metadata, "metadata is null");
         this.catalogName = requireNonNull(catalogName, "catalogName is null");
-        this.staticProvider = requireNonNull(staticProvider, "staticProvider is null");
+        this.systemTables = requireNonNull(systemTables, "systemTables is null");
+        this.systemTablesMap = systemTables.stream()
+                .collect(toImmutableMap(
+                        table -> table.getTableMetadata().getTable(),
+                        identity()));
     }
 
     @Override
     public Set<SystemTable> listSystemTables(ConnectorSession session)
     {
         // dynamic are not listed, so ony list static tables
-        return staticProvider.listSystemTables(session);
+        return systemTables;
     }
 
     @Override
     public Optional<SystemTable> getSystemTable(ConnectorSession session, SchemaTableName tableName)
     {
-        Optional<SystemTable> staticSystemTable = staticProvider.getSystemTable(session, tableName);
+        Optional<SystemTable> staticSystemTable = Optional.ofNullable(systemTablesMap.get(tableName));
         if (staticSystemTable.isPresent()) {
             return staticSystemTable;
         }
 
         // This means there is no known static table, but that doesn't mean a dynamic table must exist.
         // This node could have a different config that causes that table to not exist.
-
         if (!isCoordinatorTransaction(session)) {
             // this is a session from another coordinator, so there are no dynamic tables here for that session
             return Optional.empty();
         }
-        Optional<SystemTable> systemTable = metadata.getSystemTable(
+
+        // dynamic tables that are not SINGLE_COORDINATOR mode need to implement the
+        // PageSourceProvider interface for SystemSplit through the Connector interface
+        return metadata.getSystemTable(
                 ((FullConnectorSession) session).getSession(),
                 new QualifiedObjectName(catalogName, tableName.getSchemaName(), tableName.getTableName()));
-
-        // dynamic tables require access to the transaction and thus can only run on the current coordinator
-        if (systemTable.isPresent() && systemTable.get().getDistribution() != SINGLE_COORDINATOR) {
-            throw new TrinoException(GENERIC_INTERNAL_ERROR, "Distribution for dynamic system table must be " + SINGLE_COORDINATOR);
-        }
-
-        return systemTable;
     }
 
     private boolean isCoordinatorTransaction(ConnectorSession connectorSession)
