+++ core/trino-main/src/main/java/io/trino/connector/system/SystemConnector.java
@@ -18,6 +18,7 @@ import io.trino.node.InternalNodeManager;
 import io.trino.security.AccessControl;
 import io.trino.spi.connector.ConnectorMetadata;
 import io.trino.spi.connector.ConnectorPageSourceProvider;
+import io.trino.spi.connector.ConnectorPageSourceProviderFactory;
 import io.trino.spi.connector.ConnectorSession;
 import io.trino.spi.connector.ConnectorSplitManager;
 import io.trino.spi.connector.ConnectorTransactionHandle;
@@ -25,6 +26,7 @@ import io.trino.spi.transaction.IsolationLevel;
 import io.trino.transaction.InternalConnector;
 import io.trino.transaction.TransactionId;
 
+import java.util.Optional;
 import java.util.function.Function;
 
 import static java.util.Objects.requireNonNull;
@@ -43,7 +45,8 @@ public class SystemConnector
             SystemTablesProvider tables,
             Function<TransactionId, ConnectorTransactionHandle> transactionHandleFunction,
             AccessControl accessControl,
-            String catalogName)
+            String catalogName,
+            Optional<ConnectorPageSourceProviderFactory> pageSourceProviderFactory)
     {
         requireNonNull(currentNode, "currentNode is null");
         requireNonNull(nodeManager, "nodeManager is null");
@@ -54,7 +57,7 @@ public class SystemConnector
 
         this.metadata = new SystemTablesMetadata(tables);
         this.splitManager = new SystemSplitManager(currentNode, nodeManager, tables);
-        this.pageSourceProvider = new SystemPageSourceProvider(tables, accessControl, catalogName);
+        this.pageSourceProvider = new SystemPageSourceProvider(tables, accessControl, catalogName, pageSourceProviderFactory);
         this.transactionHandleFunction = transactionHandleFunction;
     }
