+++ core/trino-spi/src/main/java/io/trino/spi/connector/SystemColumnHandle.java
@@ -11,15 +11,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.connector.system;
+package io.trino.spi.connector;
 
-import io.trino.spi.connector.ColumnHandle;
-import io.trino.spi.connector.ColumnMetadata;
-import io.trino.spi.connector.ConnectorTableMetadata;
-
-import java.util.Map;
-
-import static com.google.common.collect.ImmutableMap.toImmutableMap;
 import static io.airlift.slice.SizeOf.estimatedSizeOf;
 import static io.airlift.slice.SizeOf.instanceSize;
 import static java.util.Objects.requireNonNull;
@@ -44,12 +37,5 @@ public record SystemColumnHandle(String columnName)
     {
         return INSTANCE_SIZE
                 + estimatedSizeOf(columnName);
-    }
-
-    public static Map<String, ColumnHandle> toSystemColumnHandles(ConnectorTableMetadata tableMetadata)
-    {
-        return tableMetadata.getColumns().stream().collect(toImmutableMap(
-                ColumnMetadata::getName,
-                column -> new SystemColumnHandle(column.getName())));
     }
 }
