+++ core/trino-spi/src/main/java/io/trino/spi/connector/SystemTable.java
@@ -20,6 +20,8 @@ import java.util.Set;
 
 /**
  * Exactly one of {@link #cursor} or {@link #pageSource} must be implemented.
+ * <p>
+ * If {@link #splitSource} is implemented, the {@link Connector}'s {@link ConnectorPageSourceProvider} must handle the {@link ConnectorSplit}s it generates.
  */
 public interface SystemTable
 {
@@ -84,6 +86,16 @@ public interface SystemTable
         return pageSource(transactionHandle, session, constraint);
     }
 
+    /**
+     * Try and create a {@link ConnectorSplitSource} for the {@link SystemTable}.
+     * <p>
+     * Implementing this method in a plugin context causes {@link SystemTable#getDistribution()} to have no impact on the actual distribution of splits.
+     * The accompanying {@link Connector}'s {@link ConnectorPageSourceProvider} must handle the {@link ConnectorSplit}s.
+     *
+     * @param connectorSession the session to use for creating the data
+     * @param constraint the constraints for the table columns (indexed from 0)
+     * @return an optional {@link ConnectorSplitSource}
+     */
     default Optional<ConnectorSplitSource> splitSource(ConnectorSession connectorSession, TupleDomain<ColumnHandle> constraint)
     {
         return Optional.empty();
